add: :art: #2 foodcardwidgetを別ファイル化 ~11:45
